Handle other annotations when inserting sign-off
The Signed-off-by line should appear immediately after any other lines of a
similar format that may be used by various tools, including Change-Id (used
by Gerrit). When deciding whether to add a newline between the last
paragraph and the signature, look for whether all lines in the last
paragraph are in this format rather than just searching for a Signed-off-by
or Acked-by line in the message.

Signed-off-by: Zane Bitter <zbitter@redhat.com>