Run commit-msg hook on edit
If the Git "commit-msg" hook exists, call it on the commit message after
editing when doing "stg edit" (including in non-interactive mode, e.g. when
running "stg edit --sign").

(This simplifies the workflow for Gerrit, by allowing the ChangeId to be
inserted with the commit-msg hook using stgit.)

Fixes bug #18806.

Signed-off-by: Zane Bitter <zbitter@redhat.com>